Added local tests for get card list request
