+++ src/test/java/co/omise/resources/CardResourceTest.java
@@ -12,21 +12,6 @@ public class CardResourceTest extends ResourceTest {
     private final String CARD_ID = "card_test_4yq6tuucl9h4erukfl0";
 
     @Test
-    public void testList() throws IOException, OmiseException {
-        ScopedList<Card> list = resource().list();
-        assertEquals(1, list.getTotal());
-        assertEquals(20, list.getLimit());
-
-        Card card = list.getData().get(0);
-        assertRequested("GET", "/customers/" + CUSTOMER_ID + "/cards", 200);
-
-        assertEquals("card", card.getObject());
-        assertEquals("card_test_4yq6tuucl9h4erukfl0", card.getId());
-        assertEquals("4242", card.getLastDigits());
-        assertEquals("Visa", card.getBrand());
-    }
-
-    @Test
     public void testDestroy() throws IOException, OmiseException {
         Card card = resource().destroy(CARD_ID);
         assertRequested("DELETE", "/customers/" + CUSTOMER_ID + "/cards/" + CARD_ID, 200);
